+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Coilpack drag cancel system.
+- Coilpack IconControl events and disabled state.
 
 ## [0.3.0] 2026-05-29